werewolf-game/requirements.txt
# AI狼人杀游戏依赖包
# 版本要求: Python 3.8+

# 核心GUI框架
PyQt6>=6.0.0

# 配置管理
PyYAML>=6.0

# 网络请求
requests>=2.25.0
aiohttp>=3.8.0

# AI服务提供商
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# 异步处理
asyncio>=3.4.0

# 日志记录
logging>=0.5.0

# 数据处理
dataclasses>=0.6.0; python_version < '3.7'
typing-extensions>=4.0.0

# 开发工具 (可选)
# black>=23.0.0
# pylint>=2.17.0
# pytest>=7.0.0
# pytest-qt>=4.0.0